requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
cryptography==42.0.5
python-multipart==0.0.9
python-dotenv==1.0.1
requests==2.32.3

streamlit==1.37.1
